Removed most of the intmax_ts, which are all MinimumSize()s.

+++ windows/button.cpp
@@ -33,7 +33,7 @@ uiWindowsControlAllDefaultsExceptDestroy(uiButton)
 // from http://msdn.microsoft.com/en-us/library/windows/desktop/dn742486.aspx#sizingandspacing
 #define buttonHeight 14
 
-static void uiButtonMinimumSize(uiWindowsControl *c, intmax_t *width, intmax_t *height)
+static void uiButtonMinimumSize(uiWindowsControl *c, int *width, int *height)
 {
 	uiButton *b = uiButton(c);
 	SIZE size;
